+++ app/src/main/java/de/baumann/browser/fragment/Fragment_settings_UI.java
@@ -9,24 +9,18 @@ import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.Bundle;
 
-import androidx.appcompat.app.AlertDialog;
 import androidx.preference.EditTextPreference;
 import androidx.preference.ListPreference;
 import androidx.preference.Preference;
 import androidx.preference.PreferenceGroup;
 import androidx.preference.PreferenceManager;
 
-import com.google.android.material.dialog.MaterialAlertDialogBuilder;
-
 import java.util.Objects;
 
 import de.baumann.browser.R;
 import de.baumann.browser.preferences.BasePreferenceFragment;
-import de.baumann.browser.unit.HelperUnit;
 
 public class Fragment_settings_UI extends BasePreferenceFragment implements SharedPreferences.OnSharedPreferenceChangeListener {
-
-    private static final int REQUEST_CODE_ASK_PERMISSIONS_4 = 1234567;
 
     @Override
     public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
@@ -35,26 +29,6 @@ public class Fragment_settings_UI extends BasePreferenceFragment implements Shar
         Context context = getContext();
         assert context != null;
         initSummary(getPreferenceScreen());
-
-        Preference settings_background = findPreference("sp_tabBackground");
-        assert settings_background != null;
-        settings_background.setOnPreferenceClickListener(preference -> {
-            if (SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                int notificationAllowed = requireActivity().checkSelfPermission(Manifest.permission.POST_NOTIFICATIONS);
-                if (notificationAllowed != PackageManager.PERMISSION_GRANTED) {
-                    MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(requireActivity());
-                    builder.setIcon(R.drawable.icon_alert);
-                    builder.setMessage(R.string.app_permission);
-                    builder.setTitle(R.string.app_permission_notification);
-                    builder.setPositiveButton(R.string.app_ok, (dialog, whichButton) -> requireActivity().requestPermissions(new String[]{Manifest.permission.POST_NOTIFICATIONS}, REQUEST_CODE_ASK_PERMISSIONS_4));
-                    builder.setNegativeButton(R.string.app_cancel, (dialog, whichButton) -> dialog.cancel());
-                    AlertDialog dialog = builder.create();
-                    dialog.show();
-                    HelperUnit.setupDialog(requireActivity(), dialog);
-                }
-            }
-            return false;
-        });
     }
 
     private void initSummary(Preference p) {